romio: add long long cast to silence printf format warnings
On 64-bit machines, long has the same size as long long. MPICH configures
MPI_OFFSET to long, but ROMIO will configure it to long long. This
raises warnings when we debug print it with the format "%lld" (or %ld).
This commit patches the warnings by casting the argument to "long long".

Many parts of ROMIO code already uses this kind of cast to silence
warnings. This partch merely extends the practice to the rest of code.

I do not enjoy this kind of patching but lack of better alternatives.
Ideally, compilers should offer "auto" format such as `%v` for at least
integer and floating types.

+++ src/mpi/romio/adio/common/flatten.c
@@ -132,13 +132,14 @@ void ADIOI_Flatten(MPI_Datatype datatype, ADIOI_Flatlist_node * flat,
                 *curr_index);
     DBG_FPRINTF(stderr, "ADIOI_Flatten:: nints %#X, nadds %#X, ntypes %#X\n", nints, nadds, ntypes);
     for (i = 0; i < nints; ++i) {
-        DBG_FPRINTF(stderr, "ADIOI_Flatten:: ints[%lld]=%#X\n", i, ints[i]);
+        DBG_FPRINTF(stderr, "ADIOI_Flatten:: ints[%lld]=%#X\n", (long long) i, ints[i]);
     }
     for (i = 0; i < nadds; ++i) {
-        DBG_FPRINTF(stderr, "ADIOI_Flatten:: adds[%lld]=" MPI_AINT_FMT_HEX_SPEC "\n", i, adds[i]);
+        DBG_FPRINTF(stderr, "ADIOI_Flatten:: adds[%lld]=" MPI_AINT_FMT_HEX_SPEC "\n",
+                    (long long) i, adds[i]);
     }
     for (i = 0; i < ntypes; ++i) {
-        DBG_FPRINTF(stderr, "ADIOI_Flatten:: types[%lld]=%#llX\n", i,
+        DBG_FPRINTF(stderr, "ADIOI_Flatten:: types[%lld]=%#llX\n", (long long) i,
                     (unsigned long long) (unsigned long) types[i]);
     }
 #endif
